Generate GraphQL Float as Swift Double
The GraphQL Spec defines Float as a signed double-precision floating point. https://spec.graphql.org/October2021/#sec-Float

+++ Sources/AnimalKingdomAPI/AnimalKingdomAPI/Sources/LocalCacheMutations/AllAnimalsLocalCacheMutation.swift
@@ -57,10 +57,10 @@ public class AllAnimalsLocalCacheMutation: LocalCacheMutation {
 
         public static var __parentType: ParentType { .Object(AnimalKingdomAPI.Bird.self) }
         public static var selections: [Selection] { [
-          .field("wingspan", Float.self),
+          .field("wingspan", Double.self),
         ] }
 
-        public var wingspan: Float {
+        public var wingspan: Double {
           get { data["wingspan"] }
           set { data["wingspan"] = newValue }
         }
