+++ Sources/AnimalKingdomAPI/AnimalKingdomAPI/Sources/Operations/AllAnimalsQuery.swift
@@ -241,11 +241,11 @@ public class AllAnimalsQuery: GraphQLQuery {
           public static var __parentType: ParentType { .Object(AnimalKingdomAPI.Height.self) }
           public static var selections: [Selection] { [
             .field("relativeSize", GraphQLEnum<RelativeSize>.self),
-            .field("centimeters", Int.self),
+            .field("centimeters", Double.self),
           ] }
 
           public var relativeSize: GraphQLEnum<RelativeSize> { data["relativeSize"] }
-          public var centimeters: Int { data["centimeters"] }
+          public var centimeters: Double { data["centimeters"] }
           public var feet: Int { data["feet"] }
           public var inches: Int? { data["inches"] }
           public var meters: Int { data["meters"] }
@@ -290,7 +290,7 @@ public class AllAnimalsQuery: GraphQLQuery {
             public var inches: Int? { data["inches"] }
             public var meters: Int { data["meters"] }
             public var relativeSize: GraphQLEnum<RelativeSize> { data["relativeSize"] }
-            public var centimeters: Int { data["centimeters"] }
+            public var centimeters: Double { data["centimeters"] }
           }
         }
       }
@@ -335,7 +335,7 @@ public class AllAnimalsQuery: GraphQLQuery {
           public var inches: Int? { data["inches"] }
           public var meters: Int { data["meters"] }
           public var relativeSize: GraphQLEnum<RelativeSize> { data["relativeSize"] }
-          public var centimeters: Int { data["centimeters"] }
+          public var centimeters: Double { data["centimeters"] }
         }
       }
 
@@ -382,10 +382,10 @@ public class AllAnimalsQuery: GraphQLQuery {
 
           public static var __parentType: ParentType { .Object(AnimalKingdomAPI.Bird.self) }
           public static var selections: [Selection] { [
-            .field("wingspan", Float.self),
+            .field("wingspan", Double.self),
           ] }
 
-          public var wingspan: Float { data["wingspan"] }
+          public var wingspan: Double { data["wingspan"] }
           public var height: Height { data["height"] }
           public var species: String { data["species"] }
           public var skinCovering: GraphQLEnum<SkinCovering>? { data["skinCovering"] }
@@ -415,7 +415,7 @@ public class AllAnimalsQuery: GraphQLQuery {
             public var inches: Int? { data["inches"] }
             public var meters: Int { data["meters"] }
             public var relativeSize: GraphQLEnum<RelativeSize> { data["relativeSize"] }
-            public var centimeters: Int { data["centimeters"] }
+            public var centimeters: Double { data["centimeters"] }
           }
         }
       }
@@ -461,7 +461,7 @@ public class AllAnimalsQuery: GraphQLQuery {
           public var inches: Int? { data["inches"] }
           public var meters: Int { data["meters"] }
           public var relativeSize: GraphQLEnum<RelativeSize> { data["relativeSize"] }
-          public var centimeters: Int { data["centimeters"] }
+          public var centimeters: Double { data["centimeters"] }
         }
       }
     }
